Roll back Hydra stuff
